requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: linux-64
_libgcc_mutex=0.1=conda_forge
_openmp_mutex=4.5=2_gnu
accelerate=0.34.2=pypi_0
aiohappyeyeballs=2.4.3=pypi_0
aiohttp=3.10.8=pypi_0
aiosignal=1.3.1=pypi_0
anyio=4.6.0=pypi_0
asttokens=3.0.0=pyhd8ed1ab_1
async-timeout=4.0.3=pypi_0
attrs=24.2.0=pypi_0
av=13.0.0=pypi_0
bitsandbytes=0.41.0=pypi_0
black=24.1.0=pypi_0
blas=2.128=openblas
blas-devel=3.9.0=28_h1ea3ea9_openblas
blessed=1.20.0=pypi_0
brotli=1.1.0=pypi_0
bzip2=1.0.8=h5eee18b_6
ca-certificates=2025.1.31=hbcca054_0
certifi=2024.8.30=pypi_0
cfgv=3.4.0=pypi_0
chardet=5.2.0=pypi_0
charset-normalizer=3.3.2=pypi_0
click=8.1.7=pypi_0
colorama=0.4.6=pypi_0
comm=0.2.2=pyhd8ed1ab_1
contourpy=1.3.0=pypi_0
cudatoolkit=11.4.1=h8ab8bb3_9
cycler=0.12.1=pypi_0
dataproperty=1.0.1=pypi_0
datasets=2.16.1=pypi_0
debugpy=1.8.10=py310hf71b8c6_0
decorator=5.1.1=pyhd8ed1ab_1
decord=0.6.0=pypi_0
deepspeed=0.14.4=pypi_0
dill=0.3.7=pypi_0
distlib=0.3.8=pypi_0
distro=1.9.0=pypi_0
docker-pycreds=0.4.0=pypi_0
docstring-parser=0.16=pypi_0
einops=0.6.1=pypi_0
einops-exts=0.0.4=pypi_0
et-xmlfile=1.1.0=pypi_0
evaluate=0.4.3=pypi_0
exceptiongroup=1.2.2=pyhd8ed1ab_1
executing=2.1.0=pyhd8ed1ab_1
faiss-gpu=1.7.4=py3.10_hc0239a3_0_cuda11.4
fastapi=0.115.0=pypi_0
filelock=3.16.1=pypi_0
flash-attn=2.6.3=pypi_0
fonttools=4.54.1=pypi_0
frozenlist=1.4.1=pypi_0
fsspec=2023.10.0=pypi_0
ftfy=6.2.3=pypi_0
gitdb=4.0.11=pypi_0
gitpython=3.1.43=pypi_0
gpustat=1.1.1=pypi_0
gradio-client=0.2.9=pypi_0
h11=0.14.0=pypi_0
hf-transfer=0.1.8=pypi_0
hjson=3.1.0=pypi_0
httpcore=0.16.3=pypi_0
httpx=0.23.3=pypi_0
huggingface-hub=0.25.1=pypi_0
icu=73.1=h6a678d5_0
identify=2.6.1=pypi_0
idna=3.10=pypi_0
imageio=2.35.1=pypi_0
importlib-metadata=8.5.0=pyha770c72_1
intel-openmp=2021.4.0=h06a4308_3561
ipykernel=6.29.5=pyh3099207_0
ipython=8.30.0=pyh707e725_0
isort=5.13.2=pypi_0
jedi=0.19.2=pyhd8ed1ab_1
jinja2=3.1.4=pypi_0
jiter=0.5.0=pypi_0
joblib=1.4.2=pypi_0
jsonlines=4.0.0=pypi_0
jupyter_client=8.6.3=pyhd8ed1ab_1
jupyter_core=5.7.2=pyh31011fe_1
kiwisolver=1.4.7=pypi_0
krb5=1.21.3=h143b758_0
latex2mathml=3.77.0=pypi_0
ld_impl_linux-64=2.40=h12ee557_0
libblas=3.9.0=28_h59b9bed_openblas
libcblas=3.9.0=28_he106b2a_openblas
libedit=3.1.20230828=h5eee18b_0
libfaiss=1.7.4=h13c3c6d_0_cuda11.4
libffi=3.4.4=h6a678d5_1
libgcc=14.2.0=h77fa898_1
libgcc-ng=14.2.0=h69a702a_1
libgfortran-ng=7.5.0=h14aa051_20
libgfortran4=7.5.0=h14aa051_20
libgfortran5=14.2.0=hf1ad2bd_2
libgomp=14.2.0=h77fa898_1
libhwloc=2.11.2=default_h0d58e46_1001
liblapack=3.9.0=28_h7ac8fdf_openblas
liblapacke=3.9.0=28_he2f377e_openblas
libopenblas=0.3.28=pthreads_h94d23a6_0
libsodium=1.0.20=h4ab18f5_0
libstdcxx=14.2.0=hc0a3c3a_1
libstdcxx-ng=11.2.0=h1234567_1
libuuid=1.41.5=h5eee18b_0
libxml2=2.13.5=hfdd30dd_0
llava=1.7.0.dev0=pypi_0
lmms-eval=0.2.4=pypi_0
loguru=0.7.2=pypi_0
lxml=5.3.0=pypi_0
markdown-it-py=3.0.0=pypi_0
markdown2=2.5.0=pypi_0
markupsafe=2.1.5=pypi_0
matplotlib=3.9.2=pypi_0
matplotlib-inline=0.1.7=pyhd8ed1ab_1
mbstrdecoder=1.1.3=pypi_0
mdurl=0.1.2=pypi_0
mkl=2021.4.0=h06a4308_640
mpmath=1.3.0=pypi_0
multidict=6.1.0=pypi_0
multiprocess=0.70.15=pypi_0
mutagen=1.47.0=pypi_0
mypy-extensions=1.0.0=pypi_0
ncurses=6.4=h6a678d5_0
nest-asyncio=1.6.0=pyhd8ed1ab_1
networkx=3.3=pypi_0
ninja=1.11.1.1=pypi_0
nltk=3.9.1=pypi_0
nodeenv=1.9.1=pypi_0
numexpr=2.10.1=pypi_0
numpy=1.26.4=py310heeff2f4_0
numpy-base=1.26.4=py310h8a23956_0
nvidia-cublas-cu12=12.1.3.1=pypi_0
nvidia-cuda-cupti-cu12=12.1.105=pypi_0
nvidia-cuda-nvrtc-cu12=12.1.105=pypi_0
nvidia-cuda-runtime-cu12=12.1.105=pypi_0
nvidia-cudnn-cu12=8.9.2.26=pypi_0
nvidia-cufft-cu12=11.0.2.54=pypi_0
nvidia-curand-cu12=10.3.2.106=pypi_0
nvidia-cusolver-cu12=11.4.5.107=pypi_0
nvidia-cusparse-cu12=12.1.0.106=pypi_0
nvidia-ml-py=12.560.30=pypi_0
nvidia-nccl-cu12=2.18.1=pypi_0
nvidia-nvjitlink-cu12=12.6.77=pypi_0
nvidia-nvtx-cu12=12.1.105=pypi_0
open-clip-torch=2.26.1=pypi_0
openai=1.51.0=pypi_0
openblas=0.3.28=pthreads_h6ec200e_0
opencv-python=4.10.0.84=pypi_0
opencv-python-headless=4.10.0.84=pypi_0
openpyxl=3.1.5=pypi_0
openssl=3.4.0=hb9d3cd8_0
packaging=24.1=pypi_0
pandas=2.2.3=pypi_0
parso=0.8.4=pyhd8ed1ab_1
pathspec=0.12.1=pypi_0
pathvalidate=3.2.1=pypi_0
peft=0.4.0=pypi_0
pexpect=4.9.0=pyhd8ed1ab_1
pickleshare=0.7.5=pyhd8ed1ab_1004
pillow=10.4.0=pypi_0
pip=24.2=py310h06a4308_0
platformdirs=4.3.6=pyhd8ed1ab_1
portalocker=2.10.1=pypi_0
pre-commit=4.0.0=pypi_0
prompt-toolkit=3.0.48=pyha770c72_1
protobuf=3.20.0=pypi_0
psutil=6.0.0=pypi_0
ptyprocess=0.7.0=pyhd8ed1ab_1
pure_eval=0.2.3=pyhd8ed1ab_1
py-cpuinfo=9.0.0=pypi_0
pyarrow=17.0.0=pypi_0
pyarrow-hotfix=0.6=pypi_0
pybind11=2.13.6=pypi_0
pycocoevalcap=1.2=pypi_0
pycocotools=2.0.8=pypi_0
pycryptodomex=3.21.0=pypi_0
pydantic=1.10.8=pypi_0
pygments=2.18.0=pyhd8ed1ab_1
pyparsing=3.1.4=pypi_0
pytablewriter=1.2.0=pypi_0
python=3.10.15=he870216_1
python-dateutil=2.9.0.post0=pyhff2d567_1
python_abi=3.10=2_cp310
pytz=2024.2=pypi_0
pyyaml=6.0.2=pypi_0
pyzmq=26.2.0=py310h71f11fc_3
readline=8.2=h5eee18b_0
regex=2024.9.11=pypi_0
requests=2.32.3=pypi_0
rfc3986=1.5.0=pypi_0
rich=13.9.1=pypi_0
sacrebleu=2.4.3=pypi_0
safetensors=0.4.5=pypi_0
scikit-learn=1.2.2=pypi_0
scipy=1.14.1=pypi_0
sentencepiece=0.1.99=pypi_0
sentry-sdk=2.15.0=pypi_0
setproctitle=1.3.3=pypi_0
setuptools=75.1.0=py310h06a4308_0
shortuuid=1.0.13=pypi_0
shtab=1.7.1=pypi_0
six=1.16.0=pypi_0
smmap=5.0.1=pypi_0
sniffio=1.3.1=pypi_0
sqlite=3.45.3=h5eee18b_0
sqlitedict=2.1.0=pypi_0
stack_data=0.6.3=pyhd8ed1ab_1
starlette=0.38.6=pypi_0
svgwrite=1.4.3=pypi_0
sympy=1.13.3=pypi_0
tabledata=1.3.3=pypi_0
tabulate=0.9.0=pypi_0
tbb=2022.0.0=hceb3a55_0
tbb-devel=2022.0.0=h1f99690_0
tcolorpy=0.1.6=pypi_0
tenacity=8.3.0=pypi_0
threadpoolctl=3.5.0=pypi_0
tiktoken=0.8.0=pypi_0
timm=1.0.9=pypi_0
tk=8.6.14=h39e8969_0
tokenizers=0.15.2=pypi_0
tomli=2.0.2=pypi_0
torch=2.1.2=pypi_0
torchvision=0.16.2=pypi_0
tornado=6.4.2=py310ha75aee5_0
tqdm=4.66.5=pypi_0
tqdm-multiprocess=0.0.11=pypi_0
traitlets=5.14.3=pyhd8ed1ab_1
transformers=4.40.0.dev0=pypi_0
transformers-stream-generator=0.0.5=pypi_0
triton=2.1.0=pypi_0
typepy=1.3.2=pypi_0
typing_extensions=4.12.2=pyha770c72_1
tyro=0.8.11=pypi_0
tzdata=2024.2=pypi_0
urllib3=1.26.20=pypi_0
uvicorn=0.31.0=pypi_0
virtualenv=20.26.6=pypi_0
wandb=0.18.3=pypi_0
wavedrom=2.0.3.post3=pypi_0
wcwidth=0.2.13=pyhd8ed1ab_1
websockets=13.1=pypi_0
wheel=0.44.0=py310h06a4308_0
xxhash=3.5.0=pypi_0
xz=5.4.6=h5eee18b_1
yacs=0.1.8=pypi_0
yarl=1.13.1=pypi_0
yt-dlp=2024.9.27=pypi_0
zeromq=4.3.5=h3b0a872_7
zipp=3.21.0=pyhd8ed1ab_1
zlib=1.2.13=h5eee18b_1
zss=1.2.0=pypi_0
zstandard=0.23.0=pypi_0
